# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для CDP сервиса",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Основной API"
        }
    ],
    "paths": {
...
        "/api/admin/checks/get": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Получение чеков для модерации",
                "description": "Защищено bearer-токеном",
                "operationId": "getChecks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "3"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": "22"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "test"
                                                    },
                                                    "surname": {
                                                        "type": "string",
                                                        "example": "test"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "test@test.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "9991231234"
                                                    },
                                                    "mall_id": {
                                                        "type": "integer",
                                                        "example": "32"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "file_link": {
                                                        "type": "string",
                                                        "example": "https://test.com/img.png"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/admin/checks/getFiltered": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Получение чеков для модерации по статусу",
                "description": "Защищено bearer-токеном",
                "operationId": "getFilteredChecks",
                "parameters": [
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Статус чека",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "test"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "3"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": "22"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "test"
                                                    },
                                                    "surname": {
                                                        "type": "string",
                                                        "example": "test"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "test@test.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "9991231234"
                                                    },
                                                    "mall_id": {
                                                        "type": "integer",
                                                        "example": "32"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "file_link": {
                                                        "type": "string",
                                                        "example": "https://test.com/img.png"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/admin/checks/confirm": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Подтверждение чека",
                "description": "Защищено bearer-токеном",
                "operationId": "confirmCheck",
                "parameters": [
                    {
                        "name": "check_id",
                        "in": "path",
                        "description": "Идентификатор чека",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "status",
                        "in": "path",
                        "description": "Статус чека",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "confirmed/declined"
                    },
                    {
                        "name": "check_sum",
                        "in": "path",
                        "description": "Количество начисляемых бонусов",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "300"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "3"
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": "22"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "test"
                                                },
                                                "surname": {
                                                    "type": "string",
                                                    "example": "test"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "test@test.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "9991231234"
                                                },
                                                "mall_id": {
                                                    "type": "integer",
                                                    "example": "32"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "file_link": {
                                                    "type": "string",
                                                    "example": "https://test.com/img.png"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}